x/fdpd/types/messages_calculation_time_test.go
package types

import (
	"testing"

	"delegationcontrol/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgCreateCalculationTime_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateCalculationTime
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateCalculationTime{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateCalculationTime{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgUpdateCalculationTime_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgUpdateCalculationTime
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgUpdateCalculationTime{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgUpdateCalculationTime{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgDeleteCalculationTime_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgDeleteCalculationTime
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgDeleteCalculationTime{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgDeleteCalculationTime{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
